INTG-1715 Feat with doc
