--- Runtime/Scripts/Data/Packets/NetworkPacket.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 900a40b6def544a4b36160a27e858a7a
-timeCreated: 1683162891
